Adding Game component
