+++ App.tsx
@@ -1,14 +1,7 @@
-import {Text, View} from 'react-native';
-import Board from './src/components/Board';
-import styles from './styles';
+import {Game} from './src/components/Game';
 
 function App(): JSX.Element {
-  return (
-    <View style={styles.container} testID="app">
-      <Text style={styles.title}>Tic-Tac-Toe!</Text>
-      <Board />
-    </View>
-  );
+  return <Game />;
 }
 
 export default App;
